\name{Eq.Ox.conc}
\alias{Eq.Ox.conc}

\title{
Equilibrium concentration of dissolved oxygen in water
}
\description{
Determines equilibrium dissolved oxygen concentration in water based on pressure and temperature. An estimate for barometric pressure can be generated by supplying the temperature and elevation (calculation by 'Barom.Press)
}
\usage{
Eq.Ox.conc(temp.C, elevation.m = NULL, 
bar.press = NULL, bar.units = NULL, 
out.DO.meas = "mg/L")
}

\arguments{
  \item{temp.C}{
    Water temperature in degrees C
}
  \item{elevation.m}{
    Elevation in meters. Default = 'NULL'. Must be NULL if bar.press takes a value.
}
  \item{bar.press}{
    Barometric pressure. Default = 'NULL'. Must be NULL if 'elevation.m' takes a value.
}
  \item{bar.units}{
    Units of barometric pressure for value supplied in bar.press. Must be NULL, "atm", "kpa", or"mmHg".
}
  \item{out.DO.meas}{
    Units of dissolved oxygen concentration 
}
}
\details{
}
\value{
Returns object of class numberic of full equilibrium dissolved oxygen concentration.
}
\references{
Gnaiger, Erich, and Hellmuth Forstner, eds. Polarographic oxygen sensors: Aquatic and physiological applications. Springer Science & Business Media, 2012.

Mechtly, E. A., 1973: The International System of Units, Physical Constants and Conversion Factors. NASA SP-7012, Second Revision, National Aeronautics and Space Administration, Washington, D.C.

https://en.wikipedia.org/wiki/Barometric_formula

http://nssdc.gsfc.nasa.gov/planetary/factsheet/earthfact.html
}
\author{
Tyler L. Moulton
}
\note{

}
\seealso{
 \code{\link{DO.saturation}}
 \code{\link{DO.unit.convert}}
 \code{\link{Barom.Press}}
}
\examples{
eqO2.1 <- Eq.Ox.conc(temp.C = 20, elevation.m = 1000)

eqO2.2 <- Eq.Ox.conc(temp.C = 20,
bar.press = 674.1, bar.units = "mmHg")

eqO2.1
eqO2.2

}
